app/requirements.txt
Flask==3.1.0
pymongo==4.12.1
python-dotenv==1.1.0